handle: re-export all of the wasi types used by handles
